% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_fns.R
\name{app_server}
\alias{app_server}
\title{Server logic for the Climbing Crag Visualization App}
\usage{
app_server(input, output, session, df)
}
\arguments{
\item{input, }{output, session Shiny server arguments}

\item{df}{A tibble containing crag data}
}
\value{
A Shiny server function
}
\description{
Server logic for the Climbing Crag Visualization App
}
