% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_animals.R
\name{complete_job_columns}
\alias{complete_job_columns}
\title{Ensure that the data retrieved from the Climb API has all the necessary
columns}
\usage{
complete_job_columns(data)
}
\arguments{
\item{data}{Output of \code{\link[=tidy_animals]{tidy_animals()}}}
}
\value{
A data frame containing all columns returned by the Climb Get Animals
By Job API
}
\description{
If fields have no data, they're not returned by the Climb API, which can lead
to missing columns if all the animals returned by a query have missing values
in that field. \code{xomplete_job_columns()} ensures that all columns are present
(and populated with \code{NA} if there is no data).
}
\examples{
\dontrun{
r <- get_animals()
dat <- tidy_animals(r)
complete_job_columns(dat)
}

}
