% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_module.R
\name{filterModuleUI}
\alias{filterModuleUI}
\title{UI for the Climbing Crag Filter Module}
\usage{
filterModuleUI(id)
}
\arguments{
\item{id}{A string identifier for the module, used to namespace UI elements.}
}
\value{
A UI element for the filter module, including checkbox inputs for grades and hardest moves.
}
\description{
This UI function creates the interactive filtering options for users to select climbing
grades and hardest moves from available options. These selections are used to filter the
climbing routes on the map.
}
\examples{
# Use in the Shiny app UI
filterModuleUI("filter_df")
}
