% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_animals.R
\name{get_animals}
\alias{get_animals}
\title{Get Animals}
\usage{
get_animals(token)
}
\arguments{
\item{token}{Access token; see \code{\link[=get_climb_token]{get_climb_token()}}}
}
\value{
Response object returned from the API call
}
\description{
Access the Get Animals API endpoint and retrieve information on all animals.
}
\examples{
\dontrun{
r <- get_animals("<token>") # Insert actual access token
content(r)
}
}
