% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_animals_by_cohort.R
\name{get_animals_by_cohort}
\alias{get_animals_by_cohort}
\title{Get Animals by Cohort}
\usage{
get_animals_by_cohort(cohort, token)
}
\arguments{
\item{cohort}{Cohort name}

\item{token}{Access token; see \code{\link[=get_climb_token]{get_climb_token()}}}
}
\value{
Response object returned from the API call
}
\description{
Access the Get Animals by Cohort API endpoint and retrieve information on all
animals.
}
\examples{
\dontrun{
# Insert actual access token
r <- get_animals_by_cohort(cohort = "5xFAD 6mo Tissue", "<token>")
content(r)
}
}
