% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_animals_by_job.R
\name{get_animals_by_job}
\alias{get_animals_by_job}
\title{Get Animals by Job}
\usage{
get_animals_by_job(job, token)
}
\arguments{
\item{job}{Job number}

\item{token}{Access token; see \code{\link[=get_climb_token]{get_climb_token()}}}
}
\value{
Response object returned from the API call
}
\description{
Access the Get Animals by Job API endpoint and retrieve information on all
animals.
}
\examples{
\dontrun{
# Insert actual access token
r <- get_animals_by_job(job = "6655", "<token>")
content(r)
}
}
