% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide_fns.R
\name{interpolate_tides}
\alias{interpolate_tides}
\title{Interpolate tide points}
\usage{
interpolate_tides(tide_data, time_step = 60)
}
\value{
A tibble with interpolated tide heights for each time step.
}
\description{
Interpolate tide points
}
\examples{
library(tibble)
tide_data <- tibble(time = c(3.05, 9.05, 15.5, 21.5), height = c(1.2, 9.5, 1.1, 9.1))
interpolate_tides(tide_data)
}
