% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{island_climbing}
\alias{island_climbing}
\title{Base Data From \emph{Guernsey Trad}}
\format{
A data.frame with 1,545 rows and 11 columns.
\describe{
\item{name_of_area}{General geographic area in Guernsey where the route is
located}
\item{latitude}{Latitude of the start of the climb}
\item{longitude}{Longitude of the start of the climb}
\item{climb_name}{Name of the route}
\item{tide_season}{One of "All", "Spring", or "Neap", being the types of
tide for which the climb is accessible.}
\item{tide_height}{Level of tide for which the climb is accessible}
\item{hours_to_be_inaccessible}{Number of hours that the climb is not
accessible due to water.}
\item{overall_grade_1}{}
\item{overall_grade_2}{}
\item{hardest_move_1}{}
\item{hardest_move_1}{}
}
}
\source{
(\emph{Guernsey Trad} book)
}
\usage{
island_climbing()
}
\description{
Route data transcribed from the \emph{Guernsey Trad} book.
}
