% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_animals_df.R
\name{make_animals_df}
\alias{make_animals_df}
\alias{tidy_animals_by_job}
\title{Convert metadata to a data frame}
\usage{
make_animals_df(data, include_params = TRUE)

tidy_animals_by_job(data, include_params = TRUE)
}
\arguments{
\item{data}{The output response object from \code{\link[=get_animals]{get_animals()}} or
\code{\link[=get_animals_by_job]{get_animals_by_job()}}}

\item{include_params}{Boolean indicating if query parameters should be added
as columns to the resulting data frame. This is useful for appending the
\code{jobNumber} to the resulting data frame.}
}
\value{
A data frame containing metadata on the animals
}
\description{
Given a response object, converts the animal metadata into a data frame. The
result should be relatively tidy if each animal has only one value for every
field, but there will be duplicated columns if animals have multiple values
for a given field. For example, if an animal is in multiple cohorts, there
will be multiple \code{cohortName} columns

Does a little extra work beyond \code{make_animals_df()} to ensure that some
expected columns are present.
}
\examples{
\dontrun{
r <- get_animals()
make_animals_df(r)
}

}
