% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_fns.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny App}
\usage{
run_app(df)
}
\arguments{
\item{df}{A tibble containing crag data}
}
\value{
Launches the Shiny App
}
\description{
Run the Shiny App
}
\examples{
 df <- data.frame(name_of_area = sample(paste0("Crag_",1:5),26,replace = T),climb_name = letters,overall_grade_1 = sample(paste0("E",1:11),26,replace=T),hardest_move_1 = sample(paste0(3:6,c("a","b","c")),26,replace = T))|>dplyr::mutate(overall_grade_2 = overall_grade_1, hardest_move_2 = hardest_move_1)
 run_app(df)
}
