% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_animals.R
\name{tidy_animals}
\alias{tidy_animals}
\title{Tidy the metadata returned by the API}
\usage{
tidy_animals(data, include_params = TRUE)
}
\arguments{
\item{data}{The output response object from \code{\link[=get_animals]{get_animals()}} or
\code{\link[=get_animals_by_job]{get_animals_by_job()}}}

\item{include_params}{Boolean indicating if query parameters should be added
as columns to the resulting data frame. This is useful for appending the
\code{jobNumber} to the resulting data frame.}
}
\value{
A tidy data frame containing metadata on the animals
}
\description{
Tidy the metadata returned by the API
}
\examples{
\dontrun{
r <- get_animals()
tidy_animals(r)
}

}
