% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climb_token.R
\name{climb_auth}
\alias{climb_auth}
\title{Authenticate to the CLIMB API}
\usage{
climb_auth(username, password)
}
\arguments{
\item{username}{CLIMB username}

\item{password}{CLIMB password}
}
\value{
POST response object
}
\description{
Authenticate to the CLIMB API
}
