% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_animals_df.R
\name{complete_job_columns}
\alias{complete_job_columns}
\title{Ensure that the data retrieved from the Climb API has all the necessary
columns}
\usage{
complete_job_columns(data)
}
\arguments{
\item{data}{Output of \code{\link[=make_animals_df]{make_animals_df()}}}
}
\value{
A data frame containing all columns returned by the Climb Get Animals
By Job API
}
\description{
If fields have no data, they're not returned by the Climb API, which can lead
to missing columns if all the animals returned by a query have missing values
in that field. \code{complete_job_columns()} ensures that all expected columns are
present (and populated with \code{NA} if there is no data).
}
\examples{
\dontrun{
r <- get_animals()
dat <- make_animals_df(r)
complete_job_columns(dat)
}

}
