% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_df.R
\name{filter_df}
\alias{filter_df}
\title{Filter a tibble by a parameter list}
\usage{
filter_df(df, ...)
}
\arguments{
\item{df}{a dataframe / tibble to filter}

\item{...}{A named list, where names are columns of the df and parameter vectors of filter variables.}
}
\value{
a filtered tibble / dataframe
}
\description{
Filter a tibble by a parameter list
}
\examples{
filter_df(mtcars, cyl = 6, gear =c(4,5))
}
