% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climb_token.R
\name{get_climb_token}
\alias{get_climb_token}
\title{Get an access token for the CLIMB API}
\usage{
get_climb_token(username, password)
}
\arguments{
\item{username}{CLIMB username}

\item{password}{CLIMB password}
}
\value{
An access token
}
\description{
Get an access token for the CLIMB API
}
